requirements.txt
langchain
python-dotenv
langchain_google_genai
langchain_community
pinecone
langchain-openai